requirements.txt
pymongo
requests
tornado
pynsq
raven
